# examples/campaign-image-studio-starter/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CanvaAuthStatus {
  connected: boolean;
  /** @nullable */
  expiresAt: string | null;
}

export type CanvaGenerationInputAspectRatio = typeof CanvaGenerationInputAspectRatio[keyof typeof CanvaGenerationInputAspectRatio];


export const CanvaGenerationInputAspectRatio = {
  square: 'square',
  landscape: 'landscape',
  portrait: 'portrait',
} as const;

export interface CanvaGenerationInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  prompt: string;
  aspectRatio: CanvaGenerationInputAspectRatio;
}

export type CanvaGenerationCreatedStatus = typeof CanvaGenerationCreatedStatus[keyof typeof CanvaGenerationCreatedStatus];


export const CanvaGenerationCreatedStatus = {
  in_progress: 'in_progress',
  success: 'success',
  failed: 'failed',
} as const;

export interface CanvaGenerationCreated {
  jobId: string;
  status: CanvaGenerationCreatedStatus;
}

export type CanvaGenerationJobStatus = typeof CanvaGenerationJobStatus[keyof typeof CanvaGenerationJobStatus];


export const CanvaGenerationJobStatus = {
  in_progress: 'in_progress',
  success: 'success',
  failed: 'failed',
} as const;

export interface CanvaGenerationJob {
  jobId: string;
  status: CanvaGenerationJobStatus;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  assetId: string | null;
  /** @nullable */
  error: string | null;
}

export interface CanvaImageImportInput {
  /** @minLength 1 */
  assetId: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  title: string;
}

export type CanvaImageImportCreatedStatus = typeof CanvaImageImportCreatedStatus[keyof typeof CanvaImageImportCreatedStatus];


export const CanvaImageImportCreatedStatus = {
  in_progress: 'in_progress',
  success: 'success',
  failed: 'failed',
} as const;

export interface CanvaImageImportCreated {
  jobId: string;
  status: CanvaImageImportCreatedStatus;
}

export type CanvaImageImportJobStatus = typeof CanvaImageImportJobStatus[keyof typeof CanvaImageImportJobStatus];


export const CanvaImageImportJobStatus = {
  in_progress: 'in_progress',
  success: 'success',
  failed: 'failed',
} as const;

export interface CanvaImageImportJob {
  jobId: string;
  status: CanvaImageImportJobStatus;
  /** @nullable */
  designUrl: string | null;
  /** @nullable */
  error: string | null;
}

export interface ApiError {
  error: string;
  /** @nullable */
  code: string | null;
}