# examples/campaign-image-studio-starter/lib/api-zod/src/generated/types/canvaAuthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CanvaAuthStatus {
  connected: boolean;
  /** @nullable */
  expiresAt: string | null;
}