# examples/campaign-image-studio-starter/lib/api-zod/src/generated/types/canvaGenerationCreated.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CanvaGenerationCreatedStatus } from './canvaGenerationCreatedStatus';

export interface CanvaGenerationCreated {
  jobId: string;
  status: CanvaGenerationCreatedStatus;
}