# examples/campaign-image-studio-starter/lib/api-zod/src/generated/types/canvaGenerationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CanvaGenerationInputAspectRatio } from './canvaGenerationInputAspectRatio';

export interface CanvaGenerationInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  prompt: string;
  aspectRatio: CanvaGenerationInputAspectRatio;
}