# examples/campaign-image-studio-starter/lib/api-zod/src/generated/types/canvaGenerationJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CanvaGenerationJobStatus } from './canvaGenerationJobStatus';

export interface CanvaGenerationJob {
  jobId: string;
  status: CanvaGenerationJobStatus;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  assetId: string | null;
  /** @nullable */
  error: string | null;
}