# examples/campaign-image-studio-starter/lib/api-zod/src/generated/types/canvaImageImportCreated.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CanvaImageImportCreatedStatus } from './canvaImageImportCreatedStatus';

export interface CanvaImageImportCreated {
  jobId: string;
  status: CanvaImageImportCreatedStatus;
}