# examples/campaign-image-studio-starter/lib/api-zod/src/generated/types/canvaImageImportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CanvaImageImportInput {
  /** @minLength 1 */
  assetId: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  title: string;
}