# examples/campaign-image-studio-starter/lib/api-zod/src/generated/types/canvaImageImportJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CanvaImageImportJobStatus } from './canvaImageImportJobStatus';

export interface CanvaImageImportJob {
  jobId: string;
  status: CanvaImageImportJobStatus;
  /** @nullable */
  designUrl: string | null;
  /** @nullable */
  error: string | null;
}